+++ moVirt/src/main/java/org/ovirt/mobile/movirt/ui/dashboard/general/DashboardGeneralFragment.java
@@ -16,10 +16,11 @@ import org.ovirt.mobile.movirt.ui.LoaderFragment;
 import org.ovirt.mobile.movirt.ui.MainActivity;
 import org.ovirt.mobile.movirt.ui.MainActivity_;
 import org.ovirt.mobile.movirt.ui.dashboard.PercentageCircleView;
-import org.ovirt.mobile.movirt.util.Cores;
-import org.ovirt.mobile.movirt.util.MemorySize;
-import org.ovirt.mobile.movirt.util.Percentage;
-import org.ovirt.mobile.movirt.util.UsageResource;
+import org.ovirt.mobile.movirt.ui.dashboard.general.resources.UtilizationResource;
+import org.ovirt.mobile.movirt.util.usage.Cores;
+import org.ovirt.mobile.movirt.util.usage.MemorySize;
+import org.ovirt.mobile.movirt.util.usage.Percentage;
+import org.ovirt.mobile.movirt.util.usage.UsageResource;
 
 import java.util.List;
