+++ moVirt/src/main/java/org/ovirt/mobile/movirt/ui/dashboard/general/resources/UtilizationResource.java
@@ -1,8 +1,8 @@
-package org.ovirt.mobile.movirt.ui.dashboard.general;
+package org.ovirt.mobile.movirt.ui.dashboard.general.resources;
 
-import org.ovirt.mobile.movirt.util.UsageResource;
+import org.ovirt.mobile.movirt.util.usage.UsageResource;
 
-class UtilizationResource {
+public class UtilizationResource {
     private UsageResource available;
     private UsageResource used;
     private UsageResource total;
