+++ moVirt/src/main/java/org/ovirt/mobile/movirt/ui/storage/StorageDomainDetailGeneralFragment.java
@@ -21,7 +21,7 @@ import org.ovirt.mobile.movirt.model.StorageDomain;
 import org.ovirt.mobile.movirt.provider.ProviderFacade;
 import org.ovirt.mobile.movirt.ui.ProgressBarResponse;
 import org.ovirt.mobile.movirt.ui.RefreshableLoaderFragment;
-import org.ovirt.mobile.movirt.util.MemorySize;
+import org.ovirt.mobile.movirt.util.usage.MemorySize;
 
 @EFragment(R.layout.fragment_storage_domain_detail_general)
 public class StorageDomainDetailGeneralFragment extends RefreshableLoaderFragment implements LoaderManager.LoaderCallbacks<Cursor> {
