+++ moVirt/src/main/java/org/ovirt/mobile/movirt/util/usage/UsageResource.java
@@ -1,8 +1,5 @@
-package org.ovirt.mobile.movirt.util;
+package org.ovirt.mobile.movirt.util.usage;
 
-/**
- * Created by suomiy on 5/6/16.
- */
 public interface UsageResource {
     long getValue();
